Ai/teacher/requirements.txt
streamlit>=1.30.0
openai>=1.10.0
flet>=0.21.0
python-dotenv>=1.0.0
